backend/requirements.txt
Flask>=2.2.3
Flask-Cors>=3.0.10
flask-praetorian>=1.4.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0